# openapi/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "version": "1.0.0",
    "title": "The Fabulous Roller Derby API",
    "description": "### Trouver toutes les informations sur les équipes et les matchs de Roller Derby en Europe. ![Roller Derby](https://images.vexels.com/media/users/3/148014/isolated/lists/4086a408e0bd425b7142641ac6cfa68e-roller-skate-shoe-sketch.png)",
    "termsOfService": "https://roller-derby.io/api/terms/",
    "contact": {
      "name": "FRDA Support",
      "url": "https://roller-derby.io/api/support",
      "email": "api@roller-derby.io"
    },
    "license": {
      "name": "GNU GPLv3",
      "url": "https://roller-derby.io/api/LICENSE"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5003",
      "description": "Serveur de mocks avec [Prism](https://stoplight.io/open-source/prism)"
    },
    {
      "url": "http://localhost:5004",
      "description": "Serveur de mocks avec [APISprout](https://github.com/danielgtaylor/apisprout)"
    },
    {
      "url": "http://localhost:4999",
      "description": "Serveur Express minimalist"
    }
  ],
  "security": [],
  "tags": [
    {
      "name": "Players",
      "description": "Toutes les opérations liées aux joueurs.euses."
    }
  ],
  "paths": {
    "/players": {
      "get": {
        "description": "### Retourne la liste des tous.tes les joureurs.euses.\n- Cette liste est filtrable et triable.\n- Un exemple de liste en ligne : [Les pétroleuses](http://roller-derby-caen.fr/les-petroleuses/)",
        "operationId": "getPlayerList",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "filter",
            "in": "query",
            "description": "Filters to apply to query. It's a stringified json object, with key/value filter separated by comma",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string"
            },
            "example": "{\"number\":75,\"name\":\"Vit Ô Bar\"}"
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort to apply to query. A stringified array with [sortProp, sortDirection]",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string"
            },
            "example": "[\"id\",\"ASC\"]"
          },
          {
            "name": "pagination",
            "in": "query",
            "description": "Pagination parameters. A stringified array with [perPage, currentPage]",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string"
            },
            "example": "[10,3]"
          }
        ],
        "responses": {
          "200": {
            "description": "Une liste de joueurs.euses.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "description": "Un.e joueur.euse de Roller Derby",
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "readOnly": true
                      },
                      "name": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 50
                      },
                      "number": {
                        "type": "integer",
                        "maximum": 9999,
                        "format": "int32"
                      },
                      "picture": {
                        "type": "string",
                        "nullable": true
                      },
                      "more": {
                        "type": "array",
                        "minItems": 0,
                        "maxItems": 5,
                        "items": {
                          "type": "object",
                          "properties": {
                            "question": {
                              "type": "string",
                              "minLength": 10,
                              "maxLength": 999
                            },
                            "answer": {
                              "type": "string",
                              "nullable": true
                            }
                          },
                          "required": [
                            "question"
                          ]
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "number",
                      "picture"
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Ajoute un.e joueur.euse.",
        "operationId": "createPlayer",
        "tags": [
          "Players"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "player": {
                    "description": "Un.e joueur.euse de Roller Derby",
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "readOnly": true
                      },
                      "name": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 50
                      },
                      "number": {
                        "type": "integer",
                        "maximum": 9999,
                        "format": "int32"
                      },
                      "picture": {
                        "type": "string",
                        "nullable": true
                      },
                      "more": {
                        "type": "array",
                        "minItems": 0,
                        "maxItems": 5,
                        "items": {
                          "type": "object",
                          "properties": {
                            "question": {
                              "type": "string",
                              "minLength": 10,
                              "maxLength": 999
                            },
                            "answer": {
                              "type": "string",
                              "nullable": true
                            }
                          },
                          "required": [
                            "question"
                          ]
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "number",
                      "picture"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Le.a joueur.euse. nouvellement créé.e.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Un.e joueur.euse de Roller Derby",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "readOnly": true
                    },
                    "name": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "number": {
                      "type": "integer",
                      "maximum": 9999,
                      "format": "int32"
                    },
                    "picture": {
                      "type": "string",
                      "nullable": true
                    },
                    "more": {
                      "type": "array",
                      "minItems": 0,
                      "maxItems": 5,
                      "items": {
                        "type": "object",
                        "properties": {
                          "question": {
                            "type": "string",
                            "minLength": 10,
                            "maxLength": 999
                          },
                          "answer": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "question"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "number",
                    "picture"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/players/{playerId}": {
      "get": {
        "description": "Obtenir toutes les informations sur un.e joueur.euse.",
        "operationId": "getPlayer",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "description": "L'identifiant du.de la joueur.euse.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Les informations sur le.a joueur.euse demandé.e.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Un.e joueur.euse de Roller Derby",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "readOnly": true
                    },
                    "name": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "number": {
                      "type": "integer",
                      "maximum": 9999,
                      "format": "int32"
                    },
                    "picture": {
                      "type": "string",
                      "nullable": true
                    },
                    "more": {
                      "type": "array",
                      "minItems": 0,
                      "maxItems": 5,
                      "items": {
                        "type": "object",
                        "properties": {
                          "question": {
                            "type": "string",
                            "minLength": 10,
                            "maxLength": 999
                          },
                          "answer": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "question"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "number",
                    "picture"
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "Mettre à jour les informations sur un.e joueur.euse.",
        "operationId": "updatePlayer",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "description": "L'identifiant du.de la joueur.euse.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "properties": {
                  "file": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  },
                  "player": {
                    "description": "Un.e joueur.euse de Roller Derby",
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "readOnly": true
                      },
                      "name": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 50
                      },
                      "number": {
                        "type": "integer",
                        "maximum": 9999,
                        "format": "int32"
                      },
                      "picture": {
                        "type": "string",
                        "nullable": true
                      },
                      "more": {
                        "type": "array",
                        "minItems": 0,
                        "maxItems": 5,
                        "items": {
                          "type": "object",
                          "properties": {
                            "question": {
                              "type": "string",
                              "minLength": 10,
                              "maxLength": 999
                            },
                            "answer": {
                              "type": "string",
                              "nullable": true
                            }
                          },
                          "required": [
                            "question"
                          ]
                        }
                      }
                    },
                    "required": [
                      "id",
                      "name",
                      "number",
                      "picture"
                    ]
                  }
                },
                "required": [
                  "player"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Les informations mises à jours du.de la joueur.euse demandé.e.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Un.e joueur.euse de Roller Derby",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "readOnly": true
                    },
                    "name": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "number": {
                      "type": "integer",
                      "maximum": 9999,
                      "format": "int32"
                    },
                    "picture": {
                      "type": "string",
                      "nullable": true
                    },
                    "more": {
                      "type": "array",
                      "minItems": 0,
                      "maxItems": 5,
                      "items": {
                        "type": "object",
                        "properties": {
                          "question": {
                            "type": "string",
                            "minLength": 10,
                            "maxLength": 999
                          },
                          "answer": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "question"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "number",
                    "picture"
                  ]
                }
              }
            }
          }
        }
      },
      "delete": {
        "description": "Supprimer un.e joueur.euse.",
        "operationId": "deletePlayer",
        "tags": [
          "Players"
        ],
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "description": "L'identifiant du.de la joueur.euse.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Les informations sur le.a joueur.euse supprimé.e.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Un.e joueur.euse de Roller Derby",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "readOnly": true
                    },
                    "name": {
                      "type": "string",
                      "minLength": 3,
                      "maxLength": 50
                    },
                    "number": {
                      "type": "integer",
                      "maximum": 9999,
                      "format": "int32"
                    },
                    "picture": {
                      "type": "string",
                      "nullable": true
                    },
                    "more": {
                      "type": "array",
                      "minItems": 0,
                      "maxItems": 5,
                      "items": {
                        "type": "object",
                        "properties": {
                          "question": {
                            "type": "string",
                            "minLength": 10,
                            "maxLength": 999
                          },
                          "answer": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "question"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "name",
                    "number",
                    "picture"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {},
    "schemas": {
      "Player": {
        "description": "Un.e joueur.euse de Roller Derby",
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 50
          },
          "number": {
            "type": "integer",
            "maximum": 9999,
            "format": "int32"
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "more": {
            "type": "array",
            "minItems": 0,
            "maxItems": 5,
            "items": {
              "type": "object",
              "properties": {
                "question": {
                  "type": "string",
                  "minLength": 10,
                  "maxLength": 999
                },
                "answer": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "question"
              ]
            }
          }
        },
        "required": [
          "id",
          "name",
          "number",
          "picture"
        ]
      }
    },
    "parameters": {}
  }
}